Script to emulate Cyanite lookups.
The script follows the same procedure as Cyanite, first querying
Elasticsearch for metric path(s) matching the glob(s) provided, then
working out the most appropriate period and rollup to search.
It then builds a Cassandra query to obtain the metric values.

A time range may be specified with the -f (from) and -t (to) flags.
Times may be timestamps or strings parseable by the Python
dateutil.parser.parse() function.  The "to" time defaults to "now" and
the "from" time defaults to 1 hour before the "to" time.

A limit of ten records will be returned by default.  Specify a different
limit with the -l flag.

Standard globs may be used.  Thus the following invocations are
equivalent, assuming there are two web servers:

  ./metrics.py servers.web1.memory.MemFree servers.web2.memory.MemFree
  ./metrics.py servers.web?.memory.MemFree
  ./metrics.py servers.web*.memory.MemFree
  ./metrics.py servers.web[12].memory.MemFree
  ./metrics.py servers.{web1,web2}.memory.MemFree