Script to find best rollup for a timestamp.
The script follows the same procedure as Cyanite for determining which
configured (period, rollup) tuple(s) is/are appropriate for a given
timestamp.  It also performs the same rounding and prints all valid
(period, rollup, time) triples.

The mandatory first argument to the script is the timestamp to process.

If no further arguments are present the script will consult the Cyanite
YAML file to determine which rollups are configured.

Alternatively the script can process one or more (period, rollup) tuples
as subsequent arguments.

Example:
  $ ./rollup.py 1431704780 1440 60 168 3600
  period:   1440; rollup:  60; time: 1431704760
  period:    168; rollup: 3600; time: 1431702000